v1.1.87

+++ models/model_shared_wallet_main_wallet_get_v3_0_response_data_main_wallet_info_unallocated_balance_local_only_unallocated_balance_available_balance.go
@@ -15,7 +15,7 @@ type SharedWalletMainWalletGetV30ResponseDataMainWalletInfoUnallocatedBalanceLoc
 	// 授信竞价可用余额(单位元)
 	CreditBiddingBalance *float64 `json:"credit_bidding_balance,omitempty"`
 	// 授信品牌可用余额(单位元)
-	CreditBrandBalance *int64 `json:"credit_brand_balance,omitempty"`
+	CreditBrandBalance *float64 `json:"credit_brand_balance,omitempty"`
 	// 授信通用可用余额(单位元)
 	CreditGeneralBalance *float64 `json:"credit_general_balance,omitempty"`
 	// 预付竞价可用余额(单位元)
